+++ Code/CUDA/cuda_fft/tsConstantGeometry.cuh
@@ -6,7 +6,7 @@
 
 __host__ void tsConstantGeometry(fftDirection dir, cpx **dev_in, cpx **dev_out, cpx *dev_W, const int n);
 
-__host__ int tsConstantGeometry_Validate(const size_t n);
-__host__ double tsConstantGeometry_Performance(const size_t n);
+__host__ int tsConstantGeometry_Validate(const int n);
+__host__ double tsConstantGeometry_Performance(const int n);
 
 #endif
