+++ Code/CUDA/cuda_fft/tsConstantGeometry_DP.cuh
@@ -6,7 +6,7 @@
 
 __host__ void tsConstantGeometry_DP(fftDirection dir, cpx **dev_in, cpx **dev_out, const int n);
 
-__host__ int tsConstantGeometry_DP_Validate(const size_t n);
-__host__ double tsConstantGeometry_DP_Performance(const size_t n);
+__host__ int tsConstantGeometry_DP_Validate(const int n);
+__host__ double tsConstantGeometry_DP_Performance(const int n);
 
 #endif
