+++ Code/CUDA/cuda_fft/tsRegular.cu
@@ -8,7 +8,7 @@
 
 __global__ void _tsRegular_body(cpx *in, cpx *out, cpx *W, const int dist, const int dist2, const int n2);
 
-__host__ int tsRegular_Validate(const size_t n)
+__host__ int tsRegular_Validate(const int n)
 {
     cpx *in, *ref, *out, *dev_in, *dev_out, *dev_W;
     fftMalloc(n, &dev_in, &dev_out, &dev_W, &in, &ref, &out);
@@ -21,7 +21,7 @@ __host__ int tsRegular_Validate(const size_t n)
     return fftResultAndFree(n, &dev_in, &dev_out, &dev_W, &in, &ref, &out) != 1;
 }
 
-__host__ double tsRegular_Performance(const size_t n)
+__host__ double tsRegular_Performance(const int n)
 {
     double measures[NUM_PERFORMANCE];
     cpx *in, *ref, *out, *dev_in, *dev_out, *dev_W;
