+++ Code/CUDA/cuda_fft/tsTobb.cuh
@@ -6,7 +6,7 @@
 
 __host__ void tsTobb(fftDirection dir, cpx **dev_in, cpx **dev_out, cpx *dev_W, const int n);
 
-__host__ int tsTobb_Validate(const size_t n);
-__host__ double tsTobb_Performance(const size_t n);
+__host__ int tsTobb_Validate(const int n);
+__host__ double tsTobb_Performance(const int n);
 
 #endif
