+++ Code/CUDA/cuda_fft/tsTobb_SB.cuh
@@ -6,7 +6,7 @@
 
 __host__ void tsTobb_SB(fftDirection dir, cpx **dev_in, cpx **dev_out, const int n);
 
-__host__ int tsTobb_SB_Validate(const size_t n);
-__host__ double tsTobb_SB_Performance(const size_t n);
+__host__ int tsTobb_SB_Validate(const int n);
+__host__ double tsTobb_SB_Performance(const int n);
 
 #endif
